requirements.txt

fastapi
uvicorn
pydantic
openai
psycopg2-binary
numpy
